add LogOut
